Part 2 of day 15 for 2015.
Added n-ingredient function to try to handle any number of ingredients.
Although it works for the supplied 2-ingredient and 4-ingredient inputs, it is
extremely slow for 4-ingredient inputs (didn't try 3 or more than 4).
Therefore, using the 4-ingredient hard-coded method for 4-ingredient inputs and
the more generic solution for anything else.